refactor: moved search to own component; made a styles file
